Refactored capi driver to move stage to its own file
